Move HybridGrant test into its respective file.
